Split all depprops into separate partial class
